Last data optimalization

+++ OpenNos.DAL.EF.MySQL/CharacterDAO.cs
@@ -60,7 +60,7 @@ namespace OpenNos.DAL.EF.MySQL
         {
             using (var context = DataAccessHelper.CreateContext())
             {
-                foreach (Character character in context.character.Where(c => c.account.Authority != (short)AuthorityType.Admin).OrderByDescending(c => c.Compliment).Take(30).ToList())
+                foreach (Character character in context.character.Where(c => c.account.Authority != (byte)AuthorityType.Admin).OrderByDescending(c => c.Compliment).Take(30).ToList())
                 {
                     yield return Mapper.Map<CharacterDTO>(character);
                 }
@@ -71,7 +71,7 @@ namespace OpenNos.DAL.EF.MySQL
         {
             using (var context = DataAccessHelper.CreateContext())
             {
-                foreach (Character character in context.character.Where(c => c.account.Authority != (short)AuthorityType.Admin).OrderByDescending(c => c.Act4Points).Take(30).ToList())
+                foreach (Character character in context.character.Where(c => c.account.Authority != (byte)AuthorityType.Admin).OrderByDescending(c => c.Act4Points).Take(30).ToList())
                 {
                     yield return Mapper.Map<CharacterDTO>(character);
                 }
@@ -82,7 +82,7 @@ namespace OpenNos.DAL.EF.MySQL
         {
             using (var context = DataAccessHelper.CreateContext())
             {
-                foreach (Character character in context.character.Where(c => c.account.Authority != (short)AuthorityType.Admin).OrderByDescending(c => c.Reput).Take(43).ToList())
+                foreach (Character character in context.character.Where(c => c.account.Authority != (byte)AuthorityType.Admin).OrderByDescending(c => c.Reput).Take(43).ToList())
                 {
                     yield return Mapper.Map<CharacterDTO>(character);
                 }
@@ -121,7 +121,7 @@ namespace OpenNos.DAL.EF.MySQL
         {
             using (var context = DataAccessHelper.CreateContext())
             {
-                List<Character> heroes = context.character.Where(c => c.account.Authority != (short)AuthorityType.Admin).OrderByDescending(c => c.Reput).Take(43).ToList();
+                List<Character> heroes = context.character.Where(c => c.account.Authority != (byte)AuthorityType.Admin).OrderByDescending(c => c.Reput).Take(43).ToList();
 
                 int i = 0;
                 foreach (Character c in heroes)
